requirements.txt
celery
Django~=5.0
djangorestframework
django-simple-history
docutils
redis
factory_boy
pytz

